Add rest of sprite 0 hit tests

+++ src/main.rs
@@ -128,7 +128,7 @@ fn main() {
 
                 if scanline < 240 {
                     window.render_scanline(&nes, scanline, &settings);
-                    nes.ppu.on_scanline(&nes.cartridge, scanline);
+                    nes.ppu.on_scanline(&nes.cartridge, scanline as i32);
                     scanline += 1;
                 } else {
                     pre_render = false;
